pyproject.toml
[build-system]
requires = ["setuptools >= 35.0.2", "wheel >= 0.29.0", "poetry>=0.12"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
force_sort_within_sections = true
forced_separate = "tests"
include_trailing_comma = true
indent = "    "
known_first_party = "seventeentrack,examples,tests"
line_length = 88
multi_line_output = 3
not_skip = "__init__.py"
sections = "FUTURE,STDLIB,INBETWEENS,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
use_parentheses = true

[tool.poetry]
name = "seventeentrack"
version = "2022.3.8"
description = "A Simple Python API for 17track.net"
readme = "README.md"
authors = ["Aaron Bach <bachya1208@gmail.com>", "James Swindle <swindle.comma.james@gmail.com>"]
license = "MIT"
repository = "https://github.com/mcswindler/seventeentrack"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

[tool.poetry.dependencies]
aiohttp = ">=3.8.0"
attrs = ">=19.3"
python = "^3.7.0"
pytz = ">=2021.1"

[tool.poetry.dev-dependencies]
aresponses = "^2.0.0"
pre-commit = "^2.0.1"
pytest = "^7.0.0"
pytest-aiohttp = "^1.0.0"
pytest-cov = "^3.0.0"
